% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_beta.R
\name{find_beta}
\alias{find_beta}
\title{Find Beta}
\usage{
find_beta(alpha, t, w, K, method = c("global", "local", "semi-param"))
}
\arguments{
\item{alpha}{A numeric between zero and one specifying the probability of false alarm.}

\item{t}{The time by which there will be this probability of a false alarm.}

\item{w}{A positive integer for the window size used by NUNC.}

\item{K}{A positive integer for the number of quantiles used by NUNC}

\item{method}{The variant of NUNC used, one of "global", "local", or "semi-parametric".}
}
\value{
A numeric for the penalty value
}
\description{
Find the threshold value for a given probability of false alarm.
}
\details{
This function provides a threshold that offers a theoretical bound on the probability
of false alarm by time t at some level alpha. In some cases, however, this bound may be 
conservative and cause a loss of test power.
}
\examples{

#Return the threshold corresponding to a 5\% probability of false alarm by time 1000 for NUNC
#local using a window of size 50 and 20 quantiles.

beta <- find_beta(alpha = 0.05, t = 1000, w = 50, K = 20, method = "local")

}
\seealso{
\link{NUNC} and \link{NUNCoffline}
}
