% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NUNCplot.R
\name{NUNCplot}
\alias{NUNCplot}
\title{NUNC Plot}
\usage{
NUNCplot(NUNCobj, xlab = NA, ylab = NA, main = NA)
}
\arguments{
\item{NUNCobj}{The output from \code{NUNCoffline}.}

\item{xlab}{The x axis label}

\item{ylab}{The y axis label}

\item{main}{The main plot title}
}
\value{
A plot of the data with vertical lines indicating the changepoint and detection time.
}
\description{
Plotting function for NUNC output
}
\details{
This function takes an output from NUNC and produces a plot of the data and, where 
applicable, the changepoint and time it is detected.
}
\examples{

#Example detection of change in distribution and subsequent plot

x <- c(rnorm(1000), rnorm(1000, 5, 3))

NUNCrun <- NUNCoffline(x, 100, 12, 15, "local")

NUNCplot(NUNCrun)


}
