% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NUNCstep.R
\name{NUNCstep}
\alias{NUNCstep}
\title{NUNC Step}
\usage{
NUNCstep(
  win,
  threshold,
  quantiles = quantile(w, probs = seq(from = 1/15, by = 1/15, length = 14)),
  method = c("global", "local"),
  params = list()
)
}
\arguments{
\item{win}{A window of data to test for a change in distribution.}

\item{threshold}{A positive numeric that gives the threshold for the test.}

\item{quantiles}{A vector of length K giving the quantiles at which the ECDFs are compared.}

\item{method}{The method used by NUNC.}

\item{params}{Additional arguments passed to NUNC - see details.}
}
\value{
Returns a \code{\link[=list]{list}} of items containing
\describe{
\item{\code{$t}}{Time change is detected - indicated by the right edge of the window.}
\item{\code{$changepoint}}{Location of the changepoint in the data stream.}
\item{\code{$method}}{Method that has been specified by the user.} 
}
}
\description{
Performs a single step of the NUNC algorithm
}
\details{
This function tests a single window of data for a change in distribution using the 
NUNC Local algorithm. This is performed by comparing the data inside the sliding window for a 
change in distribution using a cost function based on the empirical CDFs for the data.
If the difference between the CDFs is significant enough, then a change is detected.
The comparison is performed at K quantiles of the ECDF, and the information aggregated to enhance
power.

This function allows a user to specify the quantiles of the distribution, and so incorporates
both NUNC Local and NUNC Semi-Parametric

The different methods require different arguments. These arguments can be specified
through the \code{params} argument:
\describe{
\item{if \code{method = "local"}}{An extra parameter to set the maximum amount of comparisons 
 to check can be specified via extra parameter \code{max_checks}. If not specified this 
 defaults to \code{w}, the size of the window.}
}
}
\examples{

window_data <- c(rnorm(100), rnorm(100, 3, 9))
NUNCstep(window_data, threshold = 3, method = "local")

}
\seealso{
\link{NUNC} and \link{NUNCoffline}
}
